+++ packages/widgets/user-management-widget/src/lib/widget/state/asyncActions/helpers.ts
@@ -85,34 +85,5 @@ export const withRequestStatus = (
   },
 });
 
-type NotificationContent = string | { msg: string; detail?: string };
-
-const toContent = (v: NotificationContent) =>
-  typeof v === 'string' ? { msg: v } : v;
-
-export const withNotifications = ({
-  getErrorMsg,
-  getSuccessMsg,
-}: {
-  getErrorMsg?: (
-    action?: RejectedAction<AsyncThunk<any, any, any>>,
-  ) => NotificationContent;
-  getSuccessMsg?: (
-    action?: FulfilledAction<AsyncThunk<any, any, any>>,
-  ) => NotificationContent;
-}): AsyncReducerConfig<any> => ({
-  onFulfilled: (state, action) => {
-    if (getSuccessMsg)
-      state.notifications.push({
-        type: 'success',
-        ...toContent(getSuccessMsg(action)),
-      });
-  },
-  onRejected: (state, action) => {
-    if (getErrorMsg)
-      state.notifications.push({
-        type: 'error',
-        ...toContent(getErrorMsg(action)),
-      });
-  },
-});
+export { withToastNotifications as withNotifications } from '@descope/sdk-mixins';
+export type { NotificationContent } from '@descope/sdk-mixins';
